+++ src/components/PupilComponents/BrowseExamboardSelector/index.tsx
@@ -1,0 +1,1 @@
+export * from "./BrowseExamboardSelector";
