+++ src/components/PupilComponents/BrowseTierSelector/index.tsx
@@ -1,0 +1,1 @@
+export * from "./BrowseTierSelector";
